dosage amount changed to int

+++ app/src/main/java/projects/medicationtracker/SimpleClasses/Medication.java
@@ -25,7 +25,7 @@ public class Medication implements Cloneable, Parcelable {
     private LocalDateTime[] times;
     private long medId;
     private long medFrequency;
-    private float medDosage;
+    private int medDosage;
     private boolean active;
     private Medication parent = null;
     private Medication child = null;
@@ -44,7 +44,7 @@ public class Medication implements Cloneable, Parcelable {
      * @param medAlias  An alias for the Medication to appear in notifications
      */
     public Medication(String thisMed, String patient, String units, LocalDateTime[] time,
-                      LocalDateTime firstDate, long id, long frequency, float dosage, String medAlias) {
+                      LocalDateTime firstDate, long id, long frequency, int dosage, String medAlias) {
         medName = thisMed;
         patientName = patient;
         medDosageUnits = units;
@@ -230,7 +230,7 @@ public class Medication implements Cloneable, Parcelable {
      *
      * @param medDosage dosage of Medication
      */
-    public void setDosage(float medDosage) {
+    public void setDosage(int medDosage) {
         this.medDosage = medDosage;
     }
